REF: move more communication into the comm module
