+++ src/lib.rs
@@ -6,10 +6,8 @@ pub mod protocol;
 mod client;
 pub mod comm;
 mod common;
-pub mod connection;
 pub mod core;
 mod error;
-mod notifications;
 pub mod reactor;
 pub mod scheduler;
 mod task;
